requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
pymongo==4.6.1
python-dotenv==1.0.0
marshmallow==3.20.1
Werkzeug==3.0.1
gunicorn==21.2.0
